fix: update gstreamer1.0-plugins-good apt config
In the last Debian update[0], the bookworm repos changed from stable to
oldstable, which caused the original pinning config of
gstreamer1.0-plugins-good to stop working. To fix this issue and avoid
the same fix in the future, this change modifies the pinning config of
gstreamer1.0-plugins-good, so apt prefers the gstreamer1.0-plugins-good
package from Debian.

[0]: https://www.debian.org/News/2025/20250809

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>